start mac cleanup
