+++ gateware/src/tiliqua/dsp.py
@@ -220,315 +220,6 @@ class VCA(wiring.Component):
 
         return m
 
-
-#SQNative = fixed.SQ(2, ASQ.f_width)
-SQNative = fixed.SQ(2, ASQ.f_width)
-
-class FastMul(wiring.Component):
-
-    """
-    Goal: use the DSP tile in the 'fast' domain at 2x the 'sync' domain,
-    from the perspective of the 'sync' domain, we get 2x multiplies / clk.
-
-    'fast' internal pipeline latency (fifo -> fifo) = 4 clocks
-    'sync' E2E pipeline latency (valid -> valid) = 7 clocks
-    """
-
-    def __init__(self, n_rr=2, fifo_depth=4):
-        self.n_rr = n_rr
-        self.fifo_depth=fifo_depth
-        super().__init__({
-            "i": In(stream.Signature(data.ArrayLayout(
-                data.StructLayout({
-                    "a": SQNative,
-                    "b": SQNative
-                }), n_rr))),
-            "o": Out(stream.Signature(data.ArrayLayout(SQNative, n_rr))),
-        })
-
-    def elaborate(self, platform):
-        m = Module()
-
-        m.submodules.i_fifo = i_fifo = AsyncFIFOBuffered(
-            width=SQNative.as_shape().width*2*self.n_rr,
-            depth=self.fifo_depth,
-            r_domain='fast',
-            w_domain='sync'
-        )
-
-        m.submodules.o_fifo = o_fifo = AsyncFIFOBuffered(
-            width=SQNative.as_shape().width*1*self.n_rr,
-            depth=self.fifo_depth,
-            r_domain='sync',
-            w_domain='fast'
-        )
-
-        wiring.connect(m, wiring.flipped(self.i), i_fifo.w_stream)
-        wiring.connect(m, o_fifo.r_stream, wiring.flipped(self.o))
-
-        i_payload_fast = Signal.like(self.i.payload)
-        o_payload_fast = Signal.like(self.o.payload)
-        m.d.comb += i_payload_fast.as_value().eq(i_fifo.r_stream.payload)
-        m.d.comb += o_fifo.w_stream.payload.eq(o_payload_fast.as_value())
-
-        i_payload0 = Signal.like(self.i.payload)
-        i_payload1 = Signal.like(self.i.payload[0])
-        m.d.fast += i_payload0.eq(i_payload_fast)
-
-        selector = Signal()
-        m.d.fast += selector.eq(selector+1)
-
-        m.d.fast += i_payload1.eq(Mux(selector, i_payload0[0], i_payload0[1]))
-
-        o_payload1 = Signal.like(self.o.payload[0])
-        o_payloads = Signal.like(self.o.payload[0])
-
-        m.d.fast += o_payload1.eq(i_payload1.a * i_payload1.b)
-        m.d.fast += o_payloads.eq(o_payload1)
-
-        m.d.comb += o_payload_fast[0].eq(o_payload1.as_value())
-        m.d.comb += o_payload_fast[1].eq(o_payloads.as_value())
-
-        # pipeline strobe
-        r_valid = Signal()
-        w_valid = Signal()
-        m.d.comb += r_valid.eq(i_fifo.r_stream.valid & i_fifo.r_stream.ready)
-        w_valid0 = Signal()
-        w_valid1 = Signal()
-        w_valid2 = Signal()
-        m.d.fast += w_valid0.eq(r_valid)
-        m.d.fast += w_valid1.eq(w_valid0)
-        m.d.fast += w_valid2.eq(w_valid1)
-        m.d.fast += w_valid.eq(w_valid2)
-        m.d.fast += [
-            o_fifo.w_stream.valid.eq(w_valid),
-            i_fifo.r_stream.ready.eq(o_fifo.w_stream.ready),
-        ]
-
-        return m
-
-class RingMessage(data.Struct):
-
-    class Kind(enum.Enum, shape=unsigned(1)):
-        INVALID     = 0
-        MUL         = 1
-
-    class Source(enum.Enum, shape=unsigned(1)):
-        CLIENT     = 0
-        SERVER     = 1
-
-    class MulClientPayload(data.Struct):
-        a: SQNative
-        b: SQNative
-
-    class MulServerPayload(data.Struct):
-        z: SQNative
-
-    source  : Source
-    kind    : Kind
-    tag     : unsigned(4)
-    payload : data.UnionLayout({
-        "mul_client": MulClientPayload,
-        "mul_server": MulServerPayload,
-    })
-
-class RingSignature(wiring.Signature):
-    def __init__(self):
-        super().__init__({
-            "i":  In(RingMessage),
-            "o":  Out(RingMessage),
-        })
-
-class RingClient(wiring.Component):
-
-    ring:   Out(RingSignature())
-
-    i:      In(RingMessage.MulClientPayload)
-    o:      Out(RingMessage.MulServerPayload)
-
-    tag:    In(4)
-    strobe: In(1)
-    valid:  Out(1)
-
-    def elaborate(self, platform):
-        m = Module()
-
-        ring = self.ring
-
-        m.d.sync += [
-            ring.o.eq(ring.i)
-        ]
-
-        wait = Signal()
-
-        # TODO: hold message after strobe until bus free?o
-
-        with m.If((ring.i.kind == RingMessage.Kind.INVALID) & self.strobe & ~wait):
-            m.d.sync += [
-                wait.eq(1),
-                ring.o.source.eq(RingMessage.Source.CLIENT),
-                ring.o.kind.eq(RingMessage.Kind.MUL),
-                ring.o.tag.eq(self.tag),
-                ring.o.payload.mul_client.eq(self.i),
-            ]
-
-        with m.If((ring.i.kind == RingMessage.Kind.MUL) &
-                  (ring.i.source == RingMessage.Source.SERVER) &
-                  (ring.i.tag == self.tag) &
-                  wait):
-
-            m.d.comb += [
-                self.valid.eq(1),
-                self.o.eq(ring.i.payload.mul_server),
-            ]
-
-            m.d.sync += [
-                ring.o.kind.eq(RingMessage.Kind.INVALID),
-                wait.eq(0),
-            ]
-
-        return m
-
-class MAC(wiring.Component):
-
-    a: In(SQNative)
-    b: In(SQNative)
-    z: Out(SQNative)
-
-    strobe: Out(1)
-    valid: Out(1)
-
-class SimpleMAC(MAC):
-
-    def elaborate(self, platform):
-        m = Module()
-        m.d.comb += [
-            self.z.eq(self.a * self.b),
-            self.valid.eq(1),
-        ]
-        return m
-
-class RingMAC(MAC):
-
-    # TODO: add MAC members
-
-    ring: Out(RingSignature())
-    tag:  In(4)
-
-    def elaborate(self, platform):
-        m = Module()
-
-        m.submodules.ring_client = ring_client = RingClient()
-        wiring.connect(m, wiring.flipped(self.ring), ring_client.ring)
-
-        m.d.comb += [
-            ring_client.tag.eq(self.tag),
-            ring_client.i.a.eq(self.a),
-            ring_client.i.b.eq(self.b),
-            ring_client.strobe.eq(self.strobe),
-            self.z.eq(ring_client.o.z),
-            self.valid.eq(ring_client.valid),
-        ]
-
-        return m
-
-class RingMACServer(wiring.Component):
-
-    def __init__(self):
-        self.clients = []
-        super().__init__({
-            "ring": Out(RingSignature())
-        })
-
-    def add_client(self):
-        self.clients.append(RingMAC())
-        return self.clients[-1]
-
-    def elaborate(self, platform):
-        m = Module()
-
-        ring = self.ring
-
-        m.d.sync += [
-            ring.o.eq(ring.i)
-        ]
-
-        # Create the ring (TODO better ordering heuristics?)
-        """
-        wiring.connect(m, wiring.flipped(ring.o), self.clients[0].i)
-        wiring.connect(m, self.clients[-1].o, wiring.flipped(ring.i))
-        for n in range(1, len(self.clients)-1):
-            wiring.connect(m, self.clients[n].o, self.clients[n+1].i)
-        """
-        m.d.comb += [
-            self.clients[0].ring.i.eq(ring.o),
-            ring.i.eq(self.clients[-1].ring.o),
-        ]
-        for n in range(len(self.clients)-1):
-            m.d.comb += self.clients[n+1].ring.i.eq(self.clients[n].ring.o)
-
-        for n in range(len(self.clients)):
-            m.d.comb += self.clients[n].tag.eq(n)
-            setattr(m.submodules, f"client{n}", self.clients[n])
-
-        with m.If((ring.i.kind == RingMessage.Kind.MUL) &
-                  (ring.i.source == RingMessage.Source.CLIENT)):
-            # Only touch the elements we want to touch
-            m.d.sync += [
-                ring.o.source.eq(RingMessage.Source.SERVER),
-                ring.o.payload.mul_server.z.eq(
-                    ring.i.payload.mul_client.a *
-                    ring.i.payload.mul_client.b),
-            ]
-
-        return m
-
-
-# each takes a __init__(self, mac=SimpleMAC()) || RingMac(bus)
-
-class MacVCA(wiring.Component):
-
-    def __init__(self, mac=None):
-        self.mac = mac
-        super().__init__({
-            "i": In(stream.Signature(data.ArrayLayout(ASQ, 2))),
-            "o": Out(stream.Signature(data.ArrayLayout(ASQ, 1))),
-        })
-
-    def elaborate(self, platform):
-        m = Module()
-
-        if self.mac is None:
-            m.submodules.mac = self.mac = SimpleMAC()
-
-        mac = self.mac
-
-        m.d.comb += [
-            mac.a.eq(self.i.payload[0]),
-            mac.b.eq(self.i.payload[1])
-        ]
-
-        with m.FSM() as fsm:
-            with m.State('WAIT-VALID'):
-                m.d.comb += self.i.ready.eq(1),
-                with m.If(self.i.valid):
-                    m.d.comb += mac.strobe.eq(1)
-                    m.next = 'WAIT-MAC'
-            with m.State('WAIT-MAC'):
-                with m.If(mac.valid):
-                    m.d.sync += [
-                        self.o.payload.eq(mac.z)
-                    ]
-                    m.next = 'WAIT-READY'
-            with m.State('WAIT-READY'):
-                m.d.comb += [
-                    self.o.valid.eq(1),
-                ]
-                with m.If(self.o.ready):
-                    m.next = 'WAIT-VALID'
-
-        return m
-
 class GainVCA(wiring.Component):
 
     """
@@ -836,16 +527,8 @@ class SVF(wiring.Component):
         n_oversample = 2
         oversample = Signal(8)
 
-        """
-        # shared multiplier for z = a*b+c
-        mac_a = Signal(dtype)
-        mac_b = Signal(dtype)
-        mac_c = Signal(dtype)
-        mac_z = Signal(dtype)
-        m.d.comb += mac_z.eq(mac_a*mac_b + mac_c)
-        """
-
         with m.FSM() as fsm:
+
             with m.State('WAIT-VALID'):
                 m.d.comb += self.i.ready.eq(1),
                 with m.If(self.i.valid):
@@ -857,38 +540,21 @@ class SVF(wiring.Component):
                    with m.If(self.i.payload.resonance.as_value()[15] == 0):
                        m.d.sync += kQinv.eq(self.i.payload.resonance)
                    m.next = 'MAC0'
-            with m.State('MAC0'):
-                m.d.comb += [
-                    mac.a.eq(abp),
-                    mac.b.eq(kK),
-                    mac.strobe.eq(1),
-                ]
-                with m.If(mac.valid):
-                    m.d.sync += alp.eq(mac.z + alp)
-                    m.next = 'MAC1'
-            with m.State('MAC1'):
-                m.d.comb += [
-                    mac.a.eq(abp),
-                    mac.b.eq(-kQinv),
-                    mac.strobe.eq(1),
-                ]
-                with m.If(mac.valid):
-                    m.d.sync += ahp.eq(mac.z + (x - alp))
-                    m.next = 'MAC2'
-            with m.State('MAC2'):
-                m.d.comb += [
-                    mac.a.eq(ahp),
-                    mac.b.eq(kK),
-                    mac.strobe.eq(1),
-                ]
-                with m.If(mac.valid):
-                    m.d.sync += abp.eq(mac.z + abp)
-                    with m.If(oversample != n_oversample - 1):
-                        m.d.sync += oversample.eq(oversample + 1)
-                        m.next = 'MAC0'
-                    with m.Else():
-                        # FIXME: average of last N oversamples, instead of last
-                        m.next = 'WAIT-READY'
+
+            mac.state(m, 'MAC0', 'MAC1', alp, abp, kK, alp)
+
+            mac.state(m, 'MAC1', 'MAC2', ahp, abp, -kQinv, x - alp)
+
+            mac.state(m, 'MAC2', 'OVER', abp, ahp, kK, abp)
+
+            with m.State('OVER'):
+                with m.If(oversample != n_oversample - 1):
+                    m.d.sync += oversample.eq(oversample + 1)
+                    m.next = 'MAC0'
+                with m.Else():
+                    # FIXME: average of last N oversamples, instead of last
+                    m.next = 'WAIT-READY'
+
             with m.State('WAIT-READY'):
                 m.d.comb += [
                     self.o.valid.eq(1),
